support for no-args funstions
